+++ programs/Xserver/hw/xfree86/drivers/trident/trident_driver.c
@@ -28,7 +28,7 @@
  *	    Massimiliano Ghilardi, max@Linuz.sns.it, some fixes to the
  *				   clockchip programming code.
  */
-/* $XFree86: xc/programs/Xserver/hw/xfree86/drivers/trident/trident_driver.c,v 1.187tsi Exp $ */
+/* $XFree86: xc/programs/Xserver/hw/xfree86/drivers/trident/trident_driver.c,v 1.188tsi Exp $ */
 
 #include "xf1bpp.h"
 #include "xf4bpp.h"
@@ -1513,7 +1513,7 @@ TRIDENTPreInit(ScrnInfoPtr pScrn, int flags)
     	}
 
     	xf86DrvMsg(pScrn->scrnIndex,X_PROBED,"IO registers at 0x%lX\n",
-							pTrident->IOAddress);
+		   (unsigned long)pTrident->IOAddress);
     }
 
     /* Register the PCI-assigned resources. */
